!!! Somehow exposed a Heisenbug in strtok_r !!!
So I modified verify_strtok() to call itself recursively in order to test
strtok_r()'s reentrancy.  This somehow triggers the "didn't get all tokens in
such-and-such" assertion, but only for situations where the test string gets
tokenized into exactly 3 tokens.  This is a Heisenbug because of the following:

1.  This bug somehow disappears when you iterate through the var args before
    calling verify_strtok().  Try running the test with macro STRTOK_R_DEBUG
    defined to 1 vs undefined to see the difference (defining this macro calls
    verify_strtok_asdf() before calling verify_strtok(), which iterates through
    the var args).

2.  Regardless of whether or not we iterate through the var args before calling
    verify_strtok(), this bug also disappears when compiling with optimizations
    disabled (-O0 in the Makefile vs -Os).  The output of cc --version is:

        cc (Ubuntu 9.3.0-17ubuntu1~20.04) 9.3.0
        Copyright (C) 2019 Free Software Foundation, Inc.

Regardless, I am still pushing this commit upstream because strtok_r()'s actual
implementation hasn't been modified.  Hopefully this isn't a compiler bug, but
then again this wouldn't be the first compiler bug I've run into...  The code
in strtok_r.c in this commit isn't my proudest, but I think it will be helpful
in resolving this bug.  TODO: testing this with clang might give us some clues?